Power Ups
Added Invulnerability and Cannon Slowing powerups. Made minor changes to Hobo Health and Crane Projectile to make them work. Started spelling power ups with a randomly generated letter when they spawn and will add to a complete word in another script that will hopefully do something based on the word in the future.